Split multi-group benchmark classes into separate files per test type
Each benchmark class now contains only one logical test group with its own
Raw_* baseline, so Ratio and AllocRatio columns compare frameworks against
the correct Raw baseline instead of a single shared one.

12 multi-group classes split into 32 single-group classes:
- AggregateBenchmarks → Count, Sum, Avg
- ComplexQueryBenchmarks → JoinFilterPaginate, MultiJoinAggregate
- CteBenchmarks → Simple, Projection, Multi
- FilterBenchmarks → WhereActive, WhereCompound, WhereById
- InsertBenchmarks → Single, Batch
- JoinBenchmarks → Inner, ThreeTable
- PaginationBenchmarks → LimitOffset, FirstPage
- SelectBenchmarks → All, Projection
- SetOperationBenchmarks → UnionAll, Intersect, Except
- StringOpBenchmarks → Contains, StartsWith
- SubqueryBenchmarks → Exists, FilteredExists, Count, Sum
- WindowFunctionBenchmarks → RowNumber, RunningSum, Rank, Lag

6 single-group classes left unchanged.
