+++ _sessions/238-benchmark-fixes/workflow.md
@@ -10,7 +10,7 @@ issue: #238
 pr:
 session: 1
 phases-total: 4
-phases-complete: 1
+phases-complete: 2
 ## Problem Statement
 Issue #238: Benchmark failures and discrepancies from 2026-04-09 run (11e6533).
