+++ src/Quarry.Benchmarks/Benchmarks/StringContainsBenchmarks.cs
@@ -9,7 +9,7 @@ using SqlKata.Compilers;
 
 namespace Quarry.Benchmarks.Benchmarks;
 
-public class StringOpBenchmarks : BenchmarkBase
+public class StringContainsBenchmarks : BenchmarkBase
 {
     // --- Contains (LIKE '%...%') ---
 
@@ -73,90 +73,6 @@ public class StringOpBenchmarks : BenchmarkBase
         var query = new Query("users")
             .Select("UserId", "UserName", "IsActive")
             .WhereContains("UserName", "User05");
-        var compiled = SqlKataCompiler.Compile(query);
-
-        await using var cmd = Connection.CreateCommand();
-        cmd.CommandText = compiled.Sql;
-        foreach (var binding in compiled.Bindings)
-        {
-            cmd.Parameters.AddWithValue($"@p{cmd.Parameters.Count}", binding);
-        }
-        await using var reader = await cmd.ExecuteReaderAsync();
-        var results = new List<UserSummaryDto>();
-        while (await reader.ReadAsync())
-        {
-            results.Add(new UserSummaryDto
-            {
-                UserId = reader.GetInt32(0),
-                UserName = reader.GetString(1),
-                IsActive = reader.GetBoolean(2)
-            });
-        }
-        return results;
-    }
-
-    // --- StartsWith (LIKE '...%') ---
-
-    [Benchmark]
-    public async Task<List<UserSummaryDto>> Raw_StartsWith()
-    {
-        await using var cmd = Connection.CreateCommand();
-        cmd.CommandText = "SELECT UserId, UserName, IsActive FROM users WHERE UserName LIKE 'User0%'";
-        await using var reader = await cmd.ExecuteReaderAsync(CommandBehavior.SingleResult | CommandBehavior.SequentialAccess);
-        var results = new List<UserSummaryDto>();
-        while (await reader.ReadAsync())
-        {
-            results.Add(new UserSummaryDto
-            {
-                UserId = reader.GetInt32(0),
-                UserName = reader.GetString(1),
-                IsActive = reader.GetBoolean(2)
-            });
-        }
-        return results;
-    }
-
-    [Benchmark]
-    public async Task<List<UserSummaryDto>> Dapper_StartsWith()
-    {
-        return (await Connection.QueryAsync<UserSummaryDto>(
-            "SELECT UserId, UserName, IsActive FROM users WHERE UserName LIKE 'User0%'")).AsList();
-    }
-
-    [Benchmark]
-    public async Task<List<UserSummaryDto>> EfCore_StartsWith()
-    {
-        return await EfContext.Users.AsNoTracking()
-            .Where(u => u.UserName.StartsWith("User0"))
-            .Select(u => new UserSummaryDto
-            {
-                UserId = u.UserId,
-                UserName = u.UserName,
-                IsActive = u.IsActive
-            })
-            .ToListAsync();
-    }
-
-    [Benchmark]
-    public async Task<List<UserSummaryDto>> Quarry_StartsWith()
-    {
-        return await QuarryDb.Users()
-            .Where(u => u.UserName.StartsWith("User0"))
-            .Select(u => new UserSummaryDto
-            {
-                UserId = u.UserId,
-                UserName = u.UserName,
-                IsActive = u.IsActive
-            })
-            .ExecuteFetchAllAsync();
-    }
-
-    [Benchmark]
-    public async Task<List<UserSummaryDto>> SqlKata_StartsWith()
-    {
-        var query = new Query("users")
-            .Select("UserId", "UserName", "IsActive")
-            .WhereStarts("UserName", "User0");
         var compiled = SqlKataCompiler.Compile(query);
 
         await using var cmd = Connection.CreateCommand();
